requirements.txt
dj-database-url==0.4.1
Django==1.11
gunicorn==19.6.0
psycopg2==2.7.3.1
whitenoise==3.2.2
django-ckeditor
django-crispy-forms
django-widget-tweaks
sendgrid